Keep the launcher open even if encountering a error on startup
Uses a QTimer to delay setting the URI till after AliasLaunchWindow is
successfully shown. This allows set_uri to error out, show the error and
stay open so the user can use the launcher to choose a different URI.

When InvalidRequirementError is encountered while setting the URI add a
label indicating the problem to the `AliasButtonGrid`.