Don't raise InvalidRequirementError's
Previously this was a bad user experience. It would show the exception
dialog twice, once for the AliasButtonGrid and once for the DistroPicker.
The first exception didn't prevent the second because they are triggered by signals.

Now the user just see the AliasButtonGrid show the important info about
the problem and the tracebacks get sent to the logger without being raised.